scheiden van de grafiek met de content
#43 heb ik een breakdown gemaakt van hoe de grafiek eruitziet, deze heb ik nu onderscheiden met de rest van het component. Ik ben alleen aan het kijken hoe dat precies verder moet de data
