+++ src/lib/index.js
@@ -1,4 +1,7 @@
 // place files you want to import through the `$lib` alias in this folder.
+export { default as Icon } from './1_Atoms/Icon.svelte'
+export { default as DonutChart } from './2_Molecules/Donut_chart.svelte'
+
 
 export { default as Title } from './4_Templates/Title.svelte'
 export { default as AccessibillityGraph } from './4_Templates/AccessibillityGraph.svelte'
@@ -8,4 +11,3 @@ export { default as TypeGraph } from './4_Templates/TypeGraph.svelte'
 export { default as Header } from './4_Templates/Header.svelte'
 export { default as Footer } from './4_Templates/Footer.svelte'
 
-export { default as Icon } from './1_Atoms/Icon.svelte'
